Reformatting and changing
Lots of changes now that Visual Studio functions correctly. Primarily turns Updater from a inner class to a friend class.
