+++ src/widgets/CustomUpdater.cpp
@@ -1,0 +1,29 @@
+﻿/**********************************************************************
+
+  Audacity: A Digital Audio Editor
+
+  CustomUpdater.cpp
+
+  Dominic Mazzoni
+
+**********************************************************************/
+
+
+#include "CustomUpdater.h"
+
+void CustomUpdater::Update(
+   wxDC& dc, const Envelope* envelope, UpdateOutputs& allOutputs) const
+{
+   TickOutputs majorOutputs{
+      allOutputs.majorLabels, allOutputs.bits, allOutputs.box };
+
+   // SET PARAMETER IN MCUSTOM CASE
+   // Works only with major labels
+
+   int numLabel = allOutputs.majorLabels.size();
+
+   for (int i = 0; (i < numLabel) && (i <= mLength); ++i)
+      TickCustom(dc, i, mFonts.major, majorOutputs);
+
+   BoxAdjust(allOutputs);
+}
