+++ src/widgets/CustomUpdater.h
@@ -11,10 +11,9 @@
 #ifndef __AUDACITY_CUSTOM_UPDATER__
 #define __AUDACITY_CUSTOM_UPDATER__
 
-#include "Ruler.h"
 #include "Updater.h"
 
-struct Ruler::CustomUpdater : public Ruler::Updater {
+struct CustomUpdater : public Updater {
    explicit CustomUpdater(const Ruler& ruler, const ZoomInfo* z)
       : Updater{ ruler, NULL }
    {}
@@ -25,19 +24,4 @@ struct Ruler::CustomUpdater : public Ruler::Updater {
    ) const override;
 };
 
-void Ruler::CustomUpdater::Update(
-   wxDC& dc, const Envelope* envelope, UpdateOutputs& allOutputs) const
-{
-   TickOutputs majorOutputs{
-      allOutputs.majorLabels, allOutputs.bits, allOutputs.box };
-
-   // SET PARAMETER IN MCUSTOM CASE
-   // Works only with major labels
-
-   int numLabel = allOutputs.majorLabels.size();
-
-   for (int i = 0; (i < numLabel) && (i <= mLength); ++i)
-      TickCustom(dc, i, mFonts.major, majorOutputs);
-
-   BoxAdjust(allOutputs);
-}
+#endif
